+++ examples/browser-workarounds.js
@@ -4,6 +4,9 @@ var isSafari = /^((?!chrome|android).)*safari/i.test(navigator.userAgent);
 var isEdge = /Edge/.test(navigator.userAgent);
 var isOpera = !!window.opera || navigator.userAgent.indexOf('OPR/') !== -1;
 
+if (typeof MediaRecorder === 'undefined') {
+}
+
 function applyAudioWorkaround() {
     if (isSafari || isEdge) {
         if (isSafari && window.MediaRecorder !== undefined) {
@@ -11,6 +14,7 @@ function applyAudioWorkaround() {
             return;
         }
 
+        // support recording in safari 11/12
         // see https://github.com/collab-project/videojs-record/issues/295
         options.plugins.record.audioRecorderType = StereoAudioRecorder;
         options.plugins.record.audioSampleRate = 44100;
@@ -34,24 +38,4 @@ function applyScreenWorkaround() {
     if (adapter.browserDetails.browser == 'firefox') {
         adapter.browserShim.shimGetDisplayMedia(window, 'screen');
     }
-}
-
-function addStartButton() {
-    var btn = document.createElement('BUTTON');
-    var t = document.createTextNode('Show player');
-    btn.onclick = createPlayer;
-    btn.appendChild(t);
-    document.body.appendChild(btn);
-}
-
-function updateContext(opts) {
-    // Safari 11 or newer automatically suspends new AudioContext's that aren't
-    // created in response to a user-gesture, like a click or tap, so create one
-    // here (inc. the script processor)
-    var AudioContext = window.AudioContext || window.webkitAudioContext;
-    var context = new AudioContext();
-    var processor = context.createScriptProcessor(1024, 1, 1);
-
-    opts.plugins.wavesurfer.audioContext = context;
-    opts.plugins.wavesurfer.audioScriptProcessor = processor;
 }